refactor: separate signature handling into dedicated file
